Create finalprojtempv3
Here's a version of the final project I made by trying to integrate the impulse response and convolution code with the file reading script. 

I'm not 100% sure on how effective this is in finding some kind of useful insight about any kind of trend in the data, but I thought I'd post it so you all can see what it looks like, try it out yourself and give your thoughts. The code only covers the first 20 files in the folder as it is, but if you want, you can adjust the "for i =" value to select a different range of files to view